added some more temp dirs for tests

+++ openmdao.util/src/openmdao/util/test/test_filewrap.py
@@ -24,15 +24,12 @@ class TestCase(unittest.TestCase):
         os.chdir(self.tempdir)
 
     def tearDown(self):
-        # if os.path.exists(self.filename):
-        #     os.remove(self.filename)
-        # if os.path.exists(self.templatename):
-        #     os.remove(self.templatename)
         os.chdir(self.startdir)
-        try:
-            shutil.rmtree(self.tempdir)
-        except OSError:
-            pass
+        if not os.environ.get('OPENMDAO_KEEPDIR', False):
+            try:
+                shutil.rmtree(self.tempdir)
+            except OSError:
+                pass
 
     def test_templated_input(self):
